+++ apps/ui-dev/component/App.jsx
@@ -3,7 +3,11 @@ import { start } from "@elf-framework/sapa";
 import { Layout } from "./Layout";
 
 function App() {
-  return <Layout>App 23334444</Layout>;
+  return (
+    <Layout>
+      <h1>Elf Framework</h1>
+    </Layout>
+  );
 }
 
 start(App);