+++ apps/ui-dev/utils/Syntax.jsx
@@ -1,0 +1,27 @@
+import { htmlToVNode } from "@elf-framework/sapa";
+
+import "./Syntax.scss";
+
+export function Syntax(code, language = "javascript") {
+  const html = window.Prism.highlight(
+    code,
+    window.Prism.languages.javascript,
+    language
+  );
+
+  return (
+    <div
+      class="syntax"
+      style={{
+        borderRadius: 4,
+        backgroundColor: "black",
+        padding: 20,
+        boxSizing: "border-box",
+        color: "white",
+      }}
+    >
+      <div class="language">{language}</div>
+      {htmlToVNode(`<pre>${html}</pre>`)}
+    </div>
+  );
+}