+++ apps/ui-dev/CHANGELOG.md
@@ -1,5 +1,25 @@
 # ui-dev
 
+## 1.0.5
+
+### Patch Changes
+
+- update document sapa basic
+
+- Updated dependencies []:
+  - @elf-framework/ui@2.0.38
+
+## 1.0.4
+
+### Patch Changes
+
+- Support Syntax, media query
+
+- Updated dependencies []:
+  - @elf-framework/ui@2.0.37
+  - @elf-framework/sapa@2.0.37
+  - @elf-framework/design-tokens@2.0.37
+
 ## 1.0.3
 
 ### Patch Changes